requirements.txt
# Django & friends
Django==1.8.14
linaro-django-pagination==2.0.3
django-ckeditor-updated==4.4.4
django-extensions==1.6.7
django-ajax-validation==0.1.3
django-jsonfield==1.0.1
django-annoying==0.9.0
django-ajax-selects==1.4.3
django-rosetta==0.7.11
django-paging==0.2.5
django-coffeescript==0.7.2
django-filter==0.13.0
django-pylibmc==0.6.1
django-otp==0.3.5
django-push-notifications==1.4.1
django-rq==0.9.1
django-celery==3.1.17
djangorestframework==3.4.1
django-modeltranslation==0.11
django-hstore==1.4.2
python-social-auth==0.2.19
sorl-thumbnail==11.12.1b
hamlpy==0.82.2
django-polymorphic-tree==1.2.5

# DB connectors
psycopg2==2.6.2
MySQL-python==1.2.5

# Utility libraries
BeautifulSoup==3.2.1
Pillow==3.3.0
structures==1.1
requests==2.10.0
lxml==3.6.1
pytils==0.3
feedparser==5.2.1
pycurl==7.43.0
pygeoip==0.3.1
ipaddr==2.1.11
libsaas==0.4
xlrd==1.0.0
xlwt==1.1.2
setproctitle==1.1.10
qrcode==5.3
phonenumbers==7.4.5
tabulate==0.7.5
croniter==0.3.12

# Cryptography
pycrypto==2.6.1
Padding==0.4

# Cache & MQ
pylibmc==1.5.1
python-memcached==1.58

# 3d Party
plivo==0.11.1
raven==5.23.0

# Other
docutils==0.12
zeep==0.14.0
django-rest-swagger==2.0.5
pyjade==4.0.0
typing==3.5.2.2

celery==3.1.17
gunicorn==19.6.0